SupplicantHal/HostapdHal: Wait for death
When wpa_supplicant/hostapd is being terminated, ensure that we block
for a small timeout (200 ms) to ensure that the daemon is indeed dead. This
prevents a potential race when wifi/tethering is quickly toggled & the
previous instance of the daemon is not fully terminated before the next
toggle on.

Bug: 134729435
Test: Manually toggled wifi/tethering multiple times in quick
succession.
Test: Will send for full regression testing.

Signed-off-by: Alex Zhao <zzc@rock-chips.com>
Change-Id: I0cd14e22bba21a4f7fdebd8e68642bcae9739b25